Move Person interface to the types folder
