+++ ui/src/Assignments/Assignment.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2021 Ford Motor Company
+ * Copyright (c) 2022 Ford Motor Company
  * All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
@@ -15,8 +15,8 @@
  * limitations under the License.
  */
 
-import {Person} from '../People/Person';
 import moment from 'moment';
+import {Person} from 'Types/Person';
 
 export interface Assignment {
     id: number;
