+++ ui/src/Assignments/AssignmentClient.ts
@@ -18,10 +18,10 @@
 import Axios, {AxiosResponse} from 'axios';
 import {CreateAssignmentsRequest, ProductPlaceholderPair} from './CreateAssignmentRequest';
 import moment from 'moment';
-import {Person} from '../People/Person';
 import {getToken} from '../Auth/TokenProvider';
 import MatomoEvents from '../Matomo/MatomoEvents';
 import {Space} from 'Types/Space';
+import {Person} from '../Types/Person';
 
 class AssignmentClient {
