+++ ui/src/Hooks/useFetchPeople/useFetchPeople.tsx
@@ -18,8 +18,8 @@
 import {useRecoilState} from 'recoil';
 import {useCallback} from 'react';
 import {PeopleState} from 'State/PeopleState';
-import {Person} from 'People/Person';
 import PeopleClient from 'People/PeopleClient';
+import {Person} from 'Types/Person';
 
 interface UseFetchPeople {
     people: Person[];
