+++ ui/src/Products/Product.test.tsx
@@ -22,7 +22,6 @@ import ProductClient from './ProductClient';
 import TestUtils, {createDataTestId, renderWithRecoil} from '../Utils/TestUtils';
 import TestData from '../Utils/TestData';
 import ProductTagClient from '../Tags/ProductTag/ProductTagClient';
-import {Person} from '../People/Person';
 import LocationClient from '../Locations/LocationClient';
 import selectEvent from 'react-select-event';
 import moment from 'moment';
@@ -30,6 +29,7 @@ import ProductCard from './ProductCard';
 import {ViewingDateState} from '../State/ViewingDateState';
 import {IsReadOnlyState} from '../State/IsReadOnlyState';
 import {Product} from '../Types/Product';
+import {Person} from '../Types/Person';
 
 jest.mock('./ProductClient');
 jest.mock('Space/SpaceClient');
@@ -177,7 +177,7 @@ describe('Products', () => {
 
             renderProductCard({ product: productWithManyAssignments });
 
-            const expectedPersonsInOrder: Array<Person> = [
+            const expectedPersonsInOrder: Person[] = [
                 {
                     name: 'Bobby',
                     id: 900,
